GameView should now be ready.

+++ src/client/data/EntityView.java
@@ -5,10 +5,10 @@ import data.Pose;
 
 public abstract class EntityView {
     protected Pose pose;
-    protected int size;
+    protected double size;
     protected String pathToGraphic;
 
-    protected EntityView(Pose pose, int size) {
+    protected EntityView(Pose pose, double size) {
         this.pose = pose;
         this.pathToGraphic = Constants.DEFAULT_GRAPHIC_PATH;
         this.size = size;
@@ -18,7 +18,7 @@ public abstract class EntityView {
         return pose;
     }
 
-    public int getSize() {
+    public double getSize() {
         return size;
     }
