+++ src/client/data/ItemDropView.java
@@ -6,7 +6,7 @@ import data.entity.item.ItemDrop;
 public class ItemDropView extends EntityView {
     protected ItemDrop name;
 
-    protected ItemDropView(Pose pose, int size, ItemDrop name) {
+    protected ItemDropView(Pose pose, double size, ItemDrop name) {
         super(pose, size);
         this.name = name;
     }
